+++ Tests/EntityFrameworkCore.CosmosDb.TestApplication/EntityFrameworkCore.CosmosDb.TestApplication.Domain/Entities/InheritanceAssociations/StandaloneDerived.cs
@@ -3,7 +3,7 @@ using Intent.RoslynWeaver.Attributes;
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.Entities.DomainEntity", Version = "1.0")]
 
-namespace EntityFrameworkCore.CosmosDb.TestApplication.Domain.Entities
+namespace EntityFrameworkCore.CosmosDb.TestApplication.Domain.Entities.InheritanceAssociations
 {
     [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
     [DefaultIntentManaged(Mode.Fully, Targets = Targets.Properties)]