+++ Tests/EntityFrameworkCore.CosmosDb.TestApplication/EntityFrameworkCore.CosmosDb.TestApplication.Infrastructure/Persistence/Configurations/InheritanceAssociations/StandaloneDerivedConfiguration.cs
@@ -1,4 +1,5 @@
 using EntityFrameworkCore.CosmosDb.TestApplication.Domain.Entities;
+using EntityFrameworkCore.CosmosDb.TestApplication.Domain.Entities.InheritanceAssociations;
 using Intent.RoslynWeaver.Attributes;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
@@ -6,7 +7,7 @@ using Microsoft.EntityFrameworkCore.Metadata.Builders;
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.EntityFrameworkCore.EntityTypeConfiguration", Version = "1.0")]
 
-namespace EntityFrameworkCore.CosmosDb.TestApplication.Infrastructure.Persistence.Configurations
+namespace EntityFrameworkCore.CosmosDb.TestApplication.Infrastructure.Persistence.Configurations.InheritanceAssociations
 {
     public class StandaloneDerivedConfiguration : IEntityTypeConfiguration<StandaloneDerived>
     {