+++ Modules/Intent.Modules.ValueObjects/release-notes.md
@@ -1,3 +1,8 @@
-﻿### Version 4.0.0 
+﻿### Version 4.1.0
+
+- Update: Adds protected constructors for serializers to use.
+- Update: Value Objects now have a `Serialization Settings` stereotype to indicate how this Value Object should be serialized.
+
+### Version 4.0.0 
 
 - New: Upgraded Templates to use new Builder Pattern paradigm.